Save view controller

+++ Picterest/Model/PhotoModel.swift
@@ -8,8 +8,8 @@ import Foundation
 
 struct PhotoModel: Codable {
     let id: String
-    let width: Int
-    let height: Int
+    let width: Double
+    let height: Double
     let urls: Urls
 }
